# chore: optimize performance (replace he with entities)