--- a/lib/routes/aozora/newbook.js
+++ b/lib/routes/aozora/newbook.js
@@ -5,7 +5,7 @@
 
 const got = require('@/utils/got'); // get web content
 const cheerio = require('cheerio'); // html parser
-const he = require('he');
+const entities = require('entities');
 
 const base_url = 'https://www.aozora.gr.jp/index_pages/';
 module.exports = async (ctx) => {
@@ -50,7 +50,7 @@ module.exports = async (ctx) => {
         let title = '';
         let title_sub = '';
         for (let j = 0; j < title_info.length; ++j) {
-            const tmp = he.decode($(title_info[j]).html()); // should convert from escaped to unicode
+            const tmp = entities.decodeXML($(title_info[j]).html()); // should convert from escaped to unicode
             if (tmp.includes('作品名：')) {
                 title = $(title_info[j]).find('td:nth-child(2)').text();
             }