--- a/lib/routes/tencent/wechat/wemp.js
+++ b/lib/routes/tencent/wechat/wemp.js
@@ -1,5 +1,5 @@
 const got = require('@/utils/got');
-const he = require('he');
+const entities = require('entities');
 const cheerio = require('cheerio');
 const date = require('@/utils/date');
 
@@ -34,7 +34,7 @@ module.exports = async (ctx) => {
                 const matchs = /"(url|temp_url)":"([^"]+mp\.weixin\.qq\.com[^"]+)"/.exec(response);
                 let weixinLink = '';
                 if (matchs && matchs[2]) {
-                    weixinLink = he.unescape(unescape(matchs[2].replace(/\\u/g, '%u')));
+                    weixinLink = entities.decodeXML(unescape(matchs[2].replace(/\\u/g, '%u')));
                 }
 
                 const single = {