Before:
````shell
➜ ab -k -c 20 -n 250 http://127.0.0.1:1200/initium/feature/zh-hant           
This is ApacheBench, Version 2.3 <$Revision: 1879490 $>
Copyright 1996 Adam Twiss, Zeus Technology Ltd, http://www.zeustech.net/
Licensed to The Apache Software Foundation, http://www.apache.org/

Benchmarking 127.0.0.1 (be patient)
Completed 100 requests
Completed 200 requests
Finished 250 requests


Server Software:        
Server Hostname:        127.0.0.1
Server Port:            1200

Document Path:          /initium/feature/zh-hant
Document Length:        224091 bytes

Concurrency Level:      20
Time taken for tests:   42.454 seconds
Complete requests:      250
Failed requests:        0
Keep-Alive requests:    250
Total transferred:      56123724 bytes
HTML transferred:       56022750 bytes
Requests per second:    5.89 [#/sec] (mean)
Time per request:       3396.299 [ms] (mean)
Time per request:       169.815 [ms] (mean, across all concurrent requests)
Transfer rate:          1291.01 [Kbytes/sec] received

Connection Times (ms)
              min  mean[+/-sd] median   max
Connect:        0    0   0.2      0       1
Processing:   131 2773 580.9   2786    6833
Waiting:      131 2772 580.8   2786    6832
Total:        131 2773 580.7   2786    6833

Percentage of the requests served within a certain time (ms)
  50%   2786
  66%   2927
  75%   3030
  80%   3120
  90%   3341
  95%   3400
  98%   3540
  99%   3547
 100%   6833 (longest request)
````


After:
````shell
➜ ab -k -c 20 -n 250 http://127.0.0.1:1200/initium/feature/zh-hant
This is ApacheBench, Version 2.3 <$Revision: 1879490 $>
Copyright 1996 Adam Twiss, Zeus Technology Ltd, http://www.zeustech.net/
Licensed to The Apache Software Foundation, http://www.apache.org/

Benchmarking 127.0.0.1 (be patient)
Completed 100 requests
Completed 200 requests
Finished 250 requests


Server Software:        
Server Hostname:        127.0.0.1
Server Port:            1200

Document Path:          /initium/feature/zh-hant
Document Length:        224164 bytes

Concurrency Level:      20
Time taken for tests:   14.996 seconds
Complete requests:      250
Failed requests:        0
Keep-Alive requests:    250
Total transferred:      56141974 bytes
HTML transferred:       56041000 bytes
Requests per second:    16.67 [#/sec] (mean)
Time per request:       1199.655 [ms] (mean)
Time per request:       59.983 [ms] (mean, across all concurrent requests)
Transfer rate:          3656.13 [Kbytes/sec] received

Connection Times (ms)
              min  mean[+/-sd] median   max
Connect:        0    0   0.4      0       3
Processing:    61  704 374.9    650    6173
Waiting:       60  704 374.8    650    6170
Total:         61  704 374.9    650    6173

Percentage of the requests served within a certain time (ms)
  50%    650
  66%    686
  75%    710
  80%    734
  90%    877
  95%   1029
  98%   1101
  99%   1102
 100%   6173 (longest request)
````