@DIYgod 2020-12-22
`entities` should be moved from devDependencies to dependencies https://github.com/DIYgod/RSSHub/blob/18adc60d16e7027ce65a358ee614109112f4fca3/package.json#L44