@oppilate 2020-12-22
Wired. So `yarn add entities` won't add it to `dependencies` if it's already in `devDependencies`. Fixed now.