@oppilate 2020-12-22
And `parse5` used by cheerio accounts for another 1/3 processing time. Cheerio is capable of handling `htmlparser2`'s parsed dom tree, which is reportedly 5 times faster than `parse5`. But that requires one more line in each cheerio parsing call:
````javascript
// Usage as of htmlparser2 version 3:
const htmlparser2 = require('htmlparser2');
const dom = htmlparser2.parseDOM(document, options);

const $ = cheerio.load(dom);
````

I won't do it right now. Just to take a note here.