@HenryQW 2020-12-22
`htmlparser2` is the default now.

https://github.com/cheeriojs/cheerio/pull/985