@oppilate 2020-12-22
> htmlparser2 is the default now.

Well, actually it's htmlparser2 for XML and parse5 for HTML: https://github.com/cheeriojs/cheerio/issues/866#issuecomment-749618522. In RSSHub both are used, which matches what I see in the profiling result.